Add start.exe that compiles correctly with g++
