--- ASCII_TIMER/ASCII_Timer_CODE/PrintHours/print_hours.hpp
@@ -1,5 +1,0 @@
-#pragma once
-#ifndef PRINT_HOURS_HPP
-#define PRINT_HOURS_HPP
-void printHours(int hours, int lineNumber, int tensOfNumberCount, bool isHoursLessThanTen);
-#endif
