--- ASCII_TIMER/ASCII_Timer_CODE/PrintHours/tempCodeRunnerFile.cpp
@@ -1,3 +1,0 @@
- case 0:
-        print0(hours / tensOfNumberCount, lineNumber, isItTheSecondDigit);
-        break;
