--- ASCII_TIMER/ASCII_Timer_CODE/PrintMinutes/print_minutes.hpp
@@ -1,5 +1,0 @@
-#pragma once
-#ifndef PRINT_MINUTES_HPP
-#define PRINT_MINUTES_HPP
-void printMinutes(const int minutes, const int lineNumber);
-#endif
