--- ASCII_TIMER/ASCII_Timer_CODE/PrintSeconds/print_seconds.hpp
@@ -1,5 +1,0 @@
-#pragma once
-#ifndef PRINT_SECONDS_HPP
-#define PRINT_SECONDS_HPP
-void printSeconds(const int seconds, const int lineNumber);
-#endif
