+++ ASCII_TIMER/ASCII_Timer_CODE/other_functions.h
@@ -1,9 +1,9 @@
 #pragma once
-#ifndef OTHER_FUNCTIONS_HPP
-#define OTHER_FUNCTIONS_HPP
+#ifndef OTHER_FUNCTIONS_H
+#define OTHER_FUNCTIONS_H
 void screenClear();
 int tensOfNumberCount(int n);
-int *transformHoursMinutesSeconds(int n);
+int* transformHoursMinutesSeconds(int n);
 void printTimer(int hours, int minutes, int seconds, int n);
 void printTimer();
 void printEndMessage();
