+++ ASCII_TIMER/ASCII_Timer_CODE/print_minutes.h
@@ -1,0 +1,5 @@
+#pragma once
+#ifndef PRINT_MINUTES_H
+#define PRINT_MINUTES_H
+void printMinutes(const int minutes, const int lineNumber, int b, int e);
+#endif
