+++ ASCII_TIMER/ASCII_Timer_CODE/print_numbers.h
@@ -1,6 +1,6 @@
 #pragma once
-#ifndef PRINT_NUMBERS_HPP
-#define PRINT_NUMBERS_HPP
+#ifndef PRINT_NUMBERS_H
+#define PRINT_NUMBERS_H
 void print0(int num, int lineNumber, bool isItTheSecondDigit);
 void print1(int num, int lineNumber, bool isItTheSecondDigit);
 void print2(int num, int lineNumber, bool isItTheSecondDigit);
