+++ ASCII_TIMER/ASCII_Timer_CODE/print_seconds.h
@@ -1,0 +1,5 @@
+#pragma once
+#ifndef PRINT_SECONDS_H
+#define PRINT_SECONDS_H
+void printSeconds(const int seconds, const int lineNumber, int c, int f);
+#endif
